+++ src/webviews/apps/shared/components/styles/base.ts
@@ -1,0 +1,16 @@
+import { css } from '@microsoft/fast-element';
+
+export const elementBase = css`
+	:host {
+		box-sizing: border-box;
+	}
+	:host *,
+	:host *::before,
+	:host *::after {
+		box-sizing: inherit;
+	}
+
+	[hidden] {
+		display: none !important;
+	}
+`;